+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning 2.0.0](https://semver.org/spec/
 - Comment count while getting posts
 - JWT refresh tokens
 - Account Recovery (forgot password feature)
+- Skills Case Normalization
 
 ## [2.0.0] - 2026-06-29